Module Colis_constraints_common__XConstraint

type t

val true_ : t

val quantify : Colis_constraints_common.Var.t -> t -> t

val unquantify : Colis_constraints_common.Var.t -> t -> t

val set_quantified_variables : Colis_constraints_common.Var.t Batteries.Seq.t -> t -> t

Conjunction with the same literals but the new quantification block.

val add_literal : Colis_constraints_common.Literal.t -> t -> t

val add_literals_list : Colis_constraints_common.Literal.t list -> t -> t

val set_literals : Colis_constraints_common.Literal.t Batteries.Seq.t -> t -> t

Conjunction with the same quantification block but the new literals.

val and_ : t -> t -> t

val equal : t -> t -> bool

val compare : t -> t -> int

val pp : Batteries.Format.formatter -> t -> unit

val pp_as_dot : name:string -> Batteries.Format.formatter -> t -> unit

val quantified_variables : t -> Colis_constraints_common.Var.t Batteries.Seq.t

val quantified_variables_set : t -> Colis_constraints_common__.Var.Set.t

val literals : t -> Colis_constraints_common.Literal.t Batteries.Seq.t

val literals_set : t -> Colis_constraints_common.Literal.Set.t

val literals_list : t -> Colis_constraints_common.Literal.t list

val from_sets : Colis_constraints_common__.Var.Set.t -> Colis_constraints_common.Literal.Set.t -> t

val is_quantified : Colis_constraints_common.Var.t -> t -> bool