Module Colis_constraints_efficient

type sat_conj

Type of satisfiable conjunctions. The interface guarantees that all objects of this type are satisfiable.

val sat_conj_to_yojson : sat_conj -> Yojson.Safe.t

val sat_conj_of_yojson : Yojson.Safe.t -> sat_conj Ppx_deriving_yojson_runtime.error_or

val true_sat_conj : sat_conj

The empty satisfiable conjunction, trivially true.

val add_to_sat_conj : Colis_constraints_common.t -> sat_conj -> sat_conj Colis_constraints_efficient__.Dnf.t

Add a constraint of type t to a satisfiable conjunction. This yields a DNF of satisfiable conjunction. The DNF can of course be empty if the added constraints make the whole formula unsatisfiable.

val simplify : sat_conj -> sat_conj

val quantify_over : Colis_constraints_common.Var.t -> sat_conj -> sat_conj

val quantify_over_and_simplify : Colis_constraints_common.Var.t -> sat_conj -> sat_conj

val pp_sat_conj : Batteries.Format.formatter -> sat_conj -> unit

val pp_sat_conj_as_dot : name:string -> Batteries.Format.formatter -> sat_conj -> unit

val sat_conj_to_literals : sat_conj -> Colis_constraints_common.Literal.t Batteries.Seq.t

val with_shadow_variables : (unit -> 'a) -> 'a

Undocumented.